requirements.txt
python-gnupg
flask
gitpython
